+++ source/map/map_manager_add_debris.cpp
@@ -39,7 +39,7 @@ void cMapManager::m_addDebris(void)
 
             // Data parsing
             std::string   dataString       = {};
-            std::uint32_t dataStringLength = 0;
+            std::uint64_t dataStringLength = 0;
             std::uint32_t tStringNum       = 0;
             std::string   tString          = {};
 
@@ -62,7 +62,7 @@ void cMapManager::m_addDebris(void)
                 /// # database name, scale min, scale max, frequency, quest name, quest required state
                 if (dataStringLength > 4)
                 {
-                    for (std::uint32_t j = 0; j < dataStringLength; ++j)
+                    for (std::uint64_t j = 0; j < dataStringLength; ++j)
                     {
                         if (dataString[j] == ' ')
                         {
@@ -153,7 +153,7 @@ void cMapManager::m_addDebris(void)
                                     tEntity->base.scale *= scale;
 
                                     // rotation
-                                    float rotation = gRandFloatNormalized() * M_PI * 2;
+                                    float rotation = gRandFloatNormalized() * static_cast<float>(M_PI) * 2.0f;
                                     tEntity->base.rotation.x = rotation * tEntity->base.rotationAxis.x;
                                     tEntity->base.rotation.y = rotation * tEntity->base.rotationAxis.y;
                                     tEntity->base.rotation.z = rotation * tEntity->base.rotationAxis.z;
