+++ source/map/map_manager_add_events.cpp
@@ -38,7 +38,7 @@ void cMapManager::m_addEvents(void)
 
             // Data parsing
             std::string   dataString       = {};
-            std::uint32_t dataStringLength = 0;
+            std::uint64_t dataStringLength = 0;
             std::uint32_t tStringNum       = 0;
             std::string   tString          = {};
 
@@ -62,7 +62,7 @@ void cMapManager::m_addEvents(void)
                 /// # Event number, type, tile x, y, trigger tile x, y, database map name, repeat
                 if (dataStringLength > 4)
                 {
-                    for (std::uint32_t j = 0; j < dataStringLength; ++j)
+                    for (std::uint64_t j = 0; j < dataStringLength; ++j)
                     {
                         if (dataString[j] == ' ')
                         {
